requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
websockets>=12.0
twilio>=9.2.1
deepgram-sdk>=3.2.0
openai>=1.30.1
elevenlabs>=1.2.2
pydantic>=2.7.2
pydantic-settings>=2.3.1
python-dotenv>=1.0.1
redis>=5.0.4
PyMuPDF>=1.24.4
python-multipart>=0.0.9
